+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Released on Thursday, February 29 2024.
 
 - Fixed `ISourceReference` not using `HtmlToken` underneath
 - Fixed handling of replacement characters in escaped script blocks (#1174)
+- Fixed handling of empty template element in SVG foreign mode (#1176)
 - Improved pre-release version to follow semver (#1171) @georg-jung
 - Improved NuGet package with link to repository (#1168) @thompson-tomo